delete row with alert

+++ MyPlaces/MainViewController.swift
@@ -11,17 +11,20 @@ import RealmSwift
 class MainViewController: UITableViewController {
  
     var places: Results<Place>!
+    var numberOfRows = 0
     
     override func viewDidLoad() {
         super.viewDidLoad()
       
         places = realm.objects(Place.self)
+        numberOfRows = places.count
     }
 
     // MARK: - Table view data source
-
+    
+    
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return places.isEmpty ? 0 : places.count
+        return places.isEmpty ? 0 : numberOfRows
     }
 
   
@@ -42,18 +45,58 @@ class MainViewController: UITableViewController {
     }
     
     // MARK: Table view delegate
-    override func tableView(_ tableView: UITableView, editActionsForRowAt indexPath: IndexPath) -> [UITableViewRowAction]? {
-        
+    private func deleteRow(rowIndexPathAt indexPath: IndexPath) -> UIContextualAction {
         let place = places[indexPath.row]
-        let deleteAction = UITableViewRowAction(style: .default, title: "Delete") { (_, _) in
-            
-            StorageManager.deleteObject(place)
-            tableView.deleteRows(at: [indexPath], with: .automatic)
+        let action = UIContextualAction(style: .destructive, title: "Delete") { [weak self] (_, _, _) in
+            guard let self = self else {return}
+            let alert = UIAlertController(title: "Are you sure?", message: "Delete \(place.name)", preferredStyle: .alert)
+            let ok = UIAlertAction(title: "Ok", style: .default) { (_) in
+                            self.numberOfRows -= 1
+                            StorageManager.deleteObject(place)
+                            self.tableView.deleteRows(at: [indexPath], with: .automatic)
+                            self.tableView.reloadData()
+            }
+            let cancel = UIAlertAction(title: "Cancel", style: .destructive) { (_) in
+                self.tableView.reloadData()
+            }
+            alert.addAction(ok)
+            alert.addAction(cancel)
+            self.present(alert, animated: true, completion: nil)
         }
-        
-        return [deleteAction]
+        return action
     }
- 
+    
+    private func editRow(rowIndexPathAt indexPath: IndexPath) -> UIContextualAction {
+        let action = UIContextualAction(style: .normal, title: "Edit") { [weak self](_, _, _) in
+            let alert = UIAlertController(title: "Do you want to edit?", message: "You could make some editing of this stuff", preferredStyle: .alert)
+            alert.addAction(UIAlertAction(title: "Ok", style: .default))
+            self?.present(alert, animated: true)
+        }
+        return action
+    }
+    override func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
+        let edit = self.editRow(rowIndexPathAt: indexPath)
+        let delete = self.deleteRow(rowIndexPathAt: indexPath)
+        let swipe = UISwipeActionsConfiguration(actions: [delete, edit])
+        return swipe
+    }
+//    override func tableView(_ tableView: UITableView, editActionsForRowAt indexPath: IndexPath) -> [UITableViewRowAction]? {
+//
+//        let place = places[indexPath.row]
+//        let deleteAction = UITableViewRowAction(style: .default, title: "Delete") { (_, _) in
+//
+//            StorageManager.deleteObject(place)
+//            tableView.deleteRows(at: [indexPath], with: .automatic)
+//        }
+//
+//        return [deleteAction]
+//    }
+    
+    
+//    override func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
+//        <#code#>
+//    }
+//
     /*
     // MARK: - Navigation
 
@@ -67,7 +110,7 @@ class MainViewController: UITableViewController {
     @IBAction func unwindSegue(_ segue: UIStoryboardSegue) {
         
         guard let newPlaceVC = segue.source as? NewPlaceViewController else { return }
-        
+        self.numberOfRows += 1
         newPlaceVC.saveNewPlace()
         tableView.reloadData()
     }
